+++ src/app/admin-panel/integrantes/integrantes.module.ts
@@ -12,17 +12,20 @@ import { SharedModule } from "./../../shared/shared.module"
 
 import { PendingChangesGuard } from "src/app/shared/guards/pending-changes.guard";
 import { TodosIntegrantesComponent } from './todos-integrantes/todos-integrantes.component';
-import { NovoIntegranteComponent } from './novo-integrante/novo-integrante.component'
+import { NovoIntegranteComponent } from './novo-integrante/novo-integrante.component';
+import { DetalhesIntegranteComponent } from './detalhes-integrante/detalhes-integrante.component'
 
 const ROUTES: Routes = [
   { path: '', component: TodosIntegrantesComponent },
   { path: 'novo', component: NovoIntegranteComponent, canDeactivate: [PendingChangesGuard]},
+  { path: 'detalhes/:nome', component: DetalhesIntegranteComponent},
 ]
 
 @NgModule({
   declarations: [
     TodosIntegrantesComponent,
-    NovoIntegranteComponent
+    NovoIntegranteComponent,
+    DetalhesIntegranteComponent
   ],
   exports: [